Fix Phaser.Image updateTransform override
