+++ Phaser/2.x/src/dragonBones/phaser/PhaserFactory.ts
@@ -9,10 +9,10 @@
  * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
  * the Software, and to permit persons to whom the Software is furnished to do so,
  * subject to the following conditions:
- * 
+ *
  * The above copyright notice and this permission notice shall be included in all
  * copies or substantial portions of the Software.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
  * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
@@ -119,7 +119,7 @@ namespace dragonBones {
             armature;
 
             const slot = BaseObject.borrowObject(PhaserSlot);
-            const rawDisplay = new Phaser.Image(PhaserFactory._game, 0.0, 0.0, Phaser.Cache.DEFAULT);
+            const rawDisplay = new PhaserSlotDisplay(PhaserFactory._game, 0.0, 0.0, Phaser.Cache.DEFAULT);
 
             slot.init(
                 slotData, displays,
