+++ Phaser/2.x/tsconfig.json
@@ -26,7 +26,7 @@
 	"files": [
 		"./libs/phaser.d.ts",
 		"../../DragonBones/src/dragonBones/modules.ts",
-		
+
 		"../../DragonBones/src/dragonBones/core/DragonBones.ts",
 		"../../DragonBones/src/dragonBones/core/BaseObject.ts",
 
@@ -62,7 +62,7 @@
 		"../../DragonBones/src/dragonBones/animation/AnimationState.ts",
 		"../../DragonBones/src/dragonBones/animation/BaseTimelineState.ts",
 		"../../DragonBones/src/dragonBones/animation/TimelineState.ts",
-		
+
 		"../../DragonBones/src/dragonBones/event/EventObject.ts",
 		"../../DragonBones/src/dragonBones/event/IEventDispatcher.ts",
 
@@ -75,6 +75,7 @@
 		"./src/dragonBones/phaser/PhaserTextureAtlasData.ts",
 		"./src/dragonBones/phaser/PhaserArmatureDisplay.ts",
 		"./src/dragonBones/phaser/PhaserSlot.ts",
+		"./src/dragonBones/phaser/PhaserSlotDisplay.ts",
 		"./src/dragonBones/phaser/PhaserFactory.ts"
 	]
 }
